The compilation is now more strict on GNU/Linux (using g++)
The options that are specific to g++/clang are moved into a if/else statement to avoid warnings on Visual Studio

+++ Libraries/DOOMful/src/utils/Profiler.cpp
@@ -39,7 +39,7 @@ namespace Doom {
     }
 
     void Profiler::stopProfiling(const std::string& name, const int& sessionID) {
-        size_t amountSessions = 0 ;
+        int amountSessions = 0 ;
         m_mutex.lock() ;
         {
             if (m_chronos.count(name) == 1) {
